add enum, fixed UpdateUserExtra, add check, rename semestr in group, delete constructor in wrapper

+++ src/orm/SendMail.cpp
@@ -1,5 +1,6 @@
 #include "../../include/orm/SendMail.h"
 #include "../../include/orm/Wrapper.h"
+#include "../../include/orm/utils.hpp"
 #include <iostream>
 
 std::vector<SendMail> SendMail::NeedSend(SQLWrapper &db) {
@@ -33,6 +34,9 @@ int SendMail::AddInQueue(SQLWrapper &db, const std::string &recipient, const std
 }
 
 void SendMail::DeleteFromQueue(SQLWrapper &db, int id) {
+    if (check_existence("send_mail", "id", id)) {
+        throw std::length_error("ERROR: FIELD NOT FOUND ");
+    }
     db << "DELETE FROM send_mail WHERE id = " << id << ";";
     db.exec();
 }
