+++ src/orm/Session.cpp
@@ -1,7 +1,11 @@
 #include "../../include/orm/Session.h"
 #include "../../include/orm/Wrapper.h"
+#include "../../include/orm/utils.hpp"
 
 Session Session::GetById(SQLWrapper &db, int id) {
+    if (check_existence("session", "id", id)) {
+        throw std::length_error("ERROR: FIELD NOT FOUND ");
+    }
     db << "SELECT * FROM session WHERE id = '" << id << "';";
     db.exec();
     Session result(
@@ -13,12 +17,17 @@ Session Session::GetById(SQLWrapper &db, int id) {
 }
 
 void Session::DeleteSession(SQLWrapper &db, int id) {
+    if (check_existence("session", "id", id)) {
+        throw std::length_error("ERROR: FIELD NOT FOUND ");
+    }
     db << "DELETE FROM session WHERE id = '" << id << "';";
     db.exec();
 }
 
 int Session::AddSession(SQLWrapper &db, int user_id, const std::string &session, time_t time_end) {
-
+    if (check_existence("student", "id", user_id)) {
+        throw std::length_error("ERROR: FIELD NOT FOUND ");
+    }
     db << "INSERT INTO session (user_id, user_session, time_end) values ('"
        << user_id << "', '"
        << session << "', to_timestamp(" << time_end << ")) returning id;";
